Formats and instructions
This file contains all the functions that we can use to execute the different instructions based on format type. And there's a function that takes the words vector (line read from file) and decides to call which function from the formats, and each function does it's own thing. I do need to pair it with the main file though, because I need to be able to access and change the actual registers, which I can't do in my file.